+++ xenonpy/model/nn/__init__.py
@@ -2,9 +2,10 @@
 #  Use of this source code is governed by a BSD-style
 #  license that can be found in the LICENSE file.
 
-__all__ = ['wrap', 'Checker', 'Layer1d', 'RegressionTrainer', 'persist', 'BaseTrainer']
+__all__ = ['wrap', 'Checker', 'Layer1d', 'RegressionTrainer', 'persist', 'Trainer']
 
-from .checker import Checker
+from .crystal_graph_cnn import ConvLayer, CrystalGraphConvNet
 from .layer import Layer1d
 from .model_maker import Generator1d
-from .runner import BaseTrainer, RegressionTrainer, persist
+from .runner import Trainer, RegressionTrainer, persist
+from .utils.checker import Checker
