--- xenonpy/model/nn/graph/__init__.py
@@ -1,8 +1,0 @@
-#  Copyright (c) 2019. TsumiNa. All rights reserved.
-#  Use of this source code is governed by a BSD-style
-#  license that can be found in the LICENSE file.
-
-
-__all__ = ['collate_pool', 'CrystalGraphDataset', 'CrystalGraphConvNet', 'ConvLayer']
-
-from .crystal_graph_cnn import collate_pool, CrystalGraphConvNet, ConvLayer, CrystalGraphDataset
