+++ xenonpy/model/nn/utils/__init__.py
@@ -2,4 +2,6 @@
 #  Use of this source code is governed by a BSD-style
 #  license that can be found in the LICENSE file.
 
-# from ._combinator import Just, combinator
+from .extension import Validator
+from .predictor import Predictor
+from .trainer import Trainer
