+++ xenonpy/model/nn/wrap/__init__.py
@@ -3,3 +3,6 @@
 #  license that can be found in the LICENSE file.
 
 from .layer import L1
+
+from .lr_scheduler import *
+from .optimizer import *
