+++ xenonpy/utils/file_path_url.py
@@ -1,0 +1,91 @@
+#  Copyright (c) 2019. yoshida-lab. All rights reserved.
+#  Use of this source code is governed by a BSD-style
+#  license that can be found in the LICENSE file.
+
+from os import getenv
+from pathlib import Path
+
+from xenonpy._conf import __github_username__, __db_version__
+from xenonpy.utils.env import config
+
+
+def get_dataset_url(name, version=__db_version__):
+    """
+    Return url with the given file name.
+
+    Args
+    ----
+    name: str
+        binary file name.
+    version: str
+        The version of repository.
+        See Also: https://github.com/yoshida-lab/dataset/releases
+
+    Return
+    ------
+    str
+        binary file url.
+    """
+    return 'https://github.com/{0}/dataset/releases/download/v{1}/{2}.pd.xz'.format(__github_username__, version,
+                                                                                    name)
+
+
+def get_data_loc(name):
+    """Return user data location"""
+
+    scheme = ('userdata', 'usermodel')
+    if name not in scheme:
+        raise ValueError('{} not in {}'.format(name, scheme))
+    if getenv(name):
+        return str(Path(getenv(name)).expanduser())
+    return str(Path(config(name)).expanduser())
+
+
+def absolute_path(path, ignore_err=True):
+    """
+    Resolve path when path include ``~``, ``parent/here``.
+
+    Parameters
+    ----------
+    path: str
+        Path to expand.
+    ignore_err: bool
+        FileNotFoundError is raised when set to False.
+        When True, the path will be created.
+    Returns
+    -------
+    str
+        Expanded path.
+    """
+    from sys import version_info
+
+    if isinstance(path, str):
+        path = Path(path)
+
+    if version_info[1] == 5:
+        if ignore_err:
+            path.expanduser().mkdir(parents=True, exist_ok=True)
+        return str(path.expanduser().resolve())
+    return str(path.expanduser().resolve(not ignore_err))
+
+
+def get_sha256(fname):
+    """
+    Calculate file's sha256 value
+
+    Parameters
+    ----------
+    fname: str
+        File name.
+
+    Returns
+    -------
+    str
+        sha256 value.
+    """
+    from hashlib import sha256
+    hasher = sha256()
+    with open(fname, "rb") as f:
+        for chunk in iter(lambda: f.read(65536), b""):
+            hasher.update(chunk)
+    return hasher.hexdigest()
