+++ xenonpy/utils/math/product.py
